refactor(codegen): split the packed-fact index lookups out of index_get.rs
#9161 pushed index_get.rs to 2041 lines, over the 2000-line cap. Moved
foreign_packed_loop_read plus the two packed-fact lookups it sits beside
into index_get/foreign_counter.rs, re-exporting packed_f64_loop_index_parts
so expr/mod.rs's existing import still resolves.

This also reattaches a doc comment the PR had orphaned: the new function
had been inserted between packed_f64_loop_fact_for_index's doc comment and
the function itself.
