+++ changelog.d/9161-packed-clone-foreign-counter-reads.md
